Load XDG cursor theme
The cursor atlas does not look great and doesn't match the cursor theme
used by the desktop environment.

We try to load the XDG cursor theme first and if unsuccessful, we
continue with the cursor atlas.

The cursor theme can be changed by the compositor using an API.

Animated cursors are not supported right now.

Closes: #20